Add shared Google OAuth module
