+++ .gitignore
@@ -14,10 +14,8 @@ dist/
 # Credentials
 client_secret_149228795840-skuhq0abp14tkkogl0tip4t1eggl4q7d.apps.googleusercontent.com.json
 service-account.json
-gmail-credentials.json
-gmail-token.json
-calendar-credentials.json
-calendar-token.json
+credentials.json
+token.json
 
 # Audio files
 voice.wav
