+++ auth.js
@@ -1,0 +1,51 @@
+const fs = require('fs');
+const path = require('path');
+const readline = require('readline');
+const { google } = require('googleapis');
+
+const SCOPES = [
+  'https://www.googleapis.com/auth/gmail.readonly',
+  'https://www.googleapis.com/auth/gmail.modify',
+  'https://www.googleapis.com/auth/gmail.send',
+  'https://www.googleapis.com/auth/calendar',
+  'https://www.googleapis.com/auth/calendar.events'
+];
+
+const CREDENTIALS_PATH = path.join(__dirname, 'credentials.json');
+const TOKEN_PATH = path.join(__dirname, 'token.json');
+
+async function authorize() {
+  if (!fs.existsSync(CREDENTIALS_PATH)) {
+    throw new Error(
+      `Credentials file not found. Please place your OAuth client JSON at ${CREDENTIALS_PATH}.`
+    );
+  }
+
+  const credentials = JSON.parse(fs.readFileSync(CREDENTIALS_PATH, 'utf8'));
+  const { client_secret, client_id, redirect_uris } = credentials.installed || credentials.web;
+  const redirectUri = (redirect_uris && redirect_uris[0]) || 'urn:ietf:wg:oauth:2.0:oob';
+
+  const oAuth2Client = new google.auth.OAuth2(client_id, client_secret, redirectUri);
+
+  if (fs.existsSync(TOKEN_PATH)) {
+    const token = JSON.parse(fs.readFileSync(TOKEN_PATH, 'utf8'));
+    oAuth2Client.setCredentials(token);
+    return oAuth2Client;
+  }
+
+  const authUrl = oAuth2Client.generateAuthUrl({ access_type: 'offline', scope: SCOPES });
+  console.log('Please visit this URL to authorize access:\n', authUrl);
+
+  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
+  const code = await new Promise(resolve => rl.question('Enter the code from that page here: ', input => {
+    rl.close();
+    resolve(input.trim());
+  }));
+
+  const { tokens } = await oAuth2Client.getToken(code);
+  oAuth2Client.setCredentials(tokens);
+  fs.writeFileSync(TOKEN_PATH, JSON.stringify(tokens));
+  return oAuth2Client;
+}
+
+module.exports = { authorize };
